Making font an optional param
