+++ SKCSwift/Views/Utility/FlowLayout.swift
@@ -9,7 +9,11 @@ import SwiftUI
 
 // stolen from https://gist.github.com/aheze/1cbd2d36764c978b28aa20a00cb4b5b6 (an amazing implementation I wish I understood and came up with)
 struct FlowLayout: Layout {
-    var spacing = CGFloat(5)
+    let spacing: CGFloat
+    
+    init(spacing: CGFloat = 5) {
+        self.spacing = spacing
+    }
     
     func sizeThatFits(proposal: ProposedViewSize, subviews: Subviews, cache: inout ()) -> CGSize {
         let containerWidth = proposal.replacingUnspecifiedDimensions().width
@@ -30,6 +34,7 @@ struct FlowLayout: Layout {
         var currentPosition: CGPoint = .zero
         var lineHeight: CGFloat = 0
         var maxX: CGFloat = 0
+        
         for size in sizes {
             if currentPosition.x + size.width > containerWidth {
                 currentPosition.x = 0
@@ -45,24 +50,6 @@ struct FlowLayout: Layout {
         }
         
         return (result, CGSize(width: maxX, height: currentPosition.y + lineHeight))
-    }
-}
-
-struct TagModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .padding(.horizontal, 4)
-            .padding(.vertical, 2)
-            .font(.caption)
-            .background(LinearGradient(
-                gradient: Gradient(colors: [Color.accentColor, Color.accentColor.opacity(0.8)]),
-                startPoint: .leading,
-                endPoint: .trailing
-            ))
-            .foregroundColor(.white)
-            .cornerRadius(6)
-            .lineLimit(1)
-            .dynamicTypeSize(...DynamicTypeSize.medium)
     }
 }
